feat(shopping-cart-context): add quantity tracking to cart items
Added a quantity key value pair to each shopping cart item
when it is added. This allows for tracking of duplicate shopping
cart items and fixes the bug on the cart page that would
list out every single product as 1 product even if it was
a duplicate of another item already in the cart.