feat(checkout): improve accordian animation
In order to prevent sharp layout shifts
with the accordian animation, I have changed
the preview componnet logic to keep the preview
components always present in the dom. The issue was that when
the preview components were conditionally rendered
there is a brief moment where there is nothing
occupying the space of the preview component resulting
in a quick contraction and expansion which is jarring.