feat(pokemon-data): fetch, store, and filter pokemon data
Due to the limitations with the walmart dataset that I
originally seeded my database with I decided to replace
some of the walmart data with pokemon data.

The main issue with the walmart data was that the categories
of each product where too specific resulting in over
5000 unique categories with no easy way to group them
into more broad categories.

Additionally, the walmart data did not have any images
and instead of manually adding images to each product for
over 5000 products I decided to replace the walmart product
image urls with pokemon images. This will also give me the abillity
to add image modals and other functionality.