refactor(db): rework database schem to use prisma exclusively
At the start of the this project I used SQL to input all
my data and create relationships between tables. This was
fine up until this point because I had not needed to
change any of the prisma models generated with the
`db pull` command. But now I needed to be able to
change these models and continued to run into an error
when attempting to migrate the database.

Error:db error: ERROR: function uuid_generate_v4() does not exist

After much filed trouble shooting I decided to pull the product
data out of the database into a csv, re-initailize the database
through prisma and then reload the data into the database.

This resolved the issue and I can now successfully migrate
the prisma models following updates.