fix(cart-context): fix shopping cart browser refresh bug
Bug: After browser refresh, products already in the
shopping cart would not have their quantity updated
when adding the same item to the cart via the productCard
component. This was due to a bug within the ShoppingCartContext.

In the `ADD_TO_CART` case of the reducer, a check is done to see
if the product already exists in the cart and if it does then
the quantity will be updated , but the previous code was attempting
to update the quantity of the productCard data payload which does
not include a quantity property.

The fix was to update the found cart product object quantity property.